修复了用户端花费和管理端记录不一致的bug

+++ EMS_version_W/package.c
@@ -187,7 +187,7 @@ current = current->next;
 return packages;
 }
 
-int markPackageAsPickedUp(int packageId,int choice) {
+double markPackageAsPickedUp(int packageId,int choice) {
 Package* package = findPackageById(packageId);
 if (package == NULL || package->status != PACKAGE_STATUS_WAITING) {
 return 0;
